2001-07-30  Aron Helser  <helser@cs.unc.edu>
	* nmg_SurfaceMask.C (nmg_SurfaceMask): Initialization change so
	initial region mode box is empty instead of whole surface.

	* nmg_GraphicsImpl.C (positionRegionBox): Only re-calculate region
	when user is done moving box icon - communicated from UI with
	highlight mode enum changes.

	* nmg_Globals.h (RegMode): Add members to region box highlight
	mode enum.

	* globjects.c (make_region_box): Follow changes in highlight
	mode.

	* nmg_SurfaceRegion.C (VERBOSECHECK): Removed FLOW conditional.

----
	* interaction.c (doRegion): Changes highlight param sent to
	graphics, allows graphics to know when user has let go of phantom
	button.

----
Others fix comments, commit warnings.

+++ src/lib/nmBase/nmb_Decoration.C
@@ -174,7 +174,7 @@ void nmb_Decoration::addScrapeMark (PointType Top, PointType Bottom,
   // we will use this later to shift all scrape marks up by the same
   // height needed to make them all visible
   if (surface_z - Bottom[2] > scrapeMaxDepth) {
-     scrapeMaxDepth = surface_z - Bottom[2];
+     scrapeMaxDepth = (float)(surface_z - Bottom[2]);
   }
 
   // Add a scrapes indicator to the list
